Fix log showing wrong classes
